Data.Array.Accelerate.LLVM.Native.Distribution.Simple

Description

Documentation

defaultMain :: IO ()

A simple implementation of main for a Cabal setup script. This is the same as defaultMain, with added support for building libraries utilising runQ*.

simpleUserHooks :: UserHooks

Hooks that correspond to a plain instantiation of the "simple" build system.

module Distribution.Simple