Data.Array.Accelerate.LLVM.Native.Foreign

Description

Documentation

data ForeignAcc f where

Constructors

ForeignAcc :: String -> (a -> Par Native (Future b)) -> ForeignAcc (a -> b)

data ForeignExp f where

Constructors

ForeignExp :: String -> IRFun1 Native () (x -> y) -> ForeignExp (x -> y)

data LLVM target a

The LLVM monad, for executing array computations. This consists of a stack for the LLVM execution context as well as the per-execution target specific state target.

data Native

Native machine code JIT execution target

Constructors

Native

Fields

linkCache :: !LinkCache

workers :: !Workers

segmentOffset :: !Bool

liftIO :: MonadIO m => IO a -> m a

Lift a computation from the IO monad.

module Data.Array.Accelerate.LLVM.Array.Data

cloneArray :: (Shape sh, Elt e) => Array sh e -> LLVM Native (Array sh e)

Copy an array into a newly allocated array. This uses memcpy.

class Monad (Par arch) => Async arch where

Minimal complete definition

new, put, get, fork

Associated Types

data Par arch a :: Type

The monad parallel computations will be executed in. Presumably a stack with the LLVM monad at the base.

type FutureR arch :: Type -> Type

Parallel computations can communicate via futures.

Methods

new :: Par arch (FutureR arch a)

Create a new (empty) promise, to be fulfilled at some future point.

put :: FutureR arch a -> a -> Par arch ()

The future is here. Multiple puts to the same future are not allowed and (presumably) result in a runtime error.

get :: FutureR arch a -> Par arch a

Read the value stored in a future, once it is available. It is _not_ required that this is a blocking operation on the host, only that it is blocking with respect to computations on the remote device.

fork :: Par arch () -> Par arch ()

Fork a computation to happen in parallel. The forked computation may exchange values with other computations using Futures.

block :: FutureR arch a -> Par arch a

Read a value stored in a future, once it is available. This is blocking with respect to both the host and remote device.

spawn :: Par arch (FutureR arch a) -> Par arch (FutureR arch a)

Evaluate a computation in a new thread/context. This might be implemented more efficiently than the default implementation.

newFull :: a -> Par arch (FutureR arch a)

Create a new "future" where the value is available immediately. This might be implemented more efficiently than the default implementation.

data Future a

Constructors

Future !(IORef (IVar a))

data IVar a

Constructors

Full !a

Blocked !(Seq (a -> IO ()))

Empty

evalPar :: Par Native a -> LLVM Native a

Evaluate a parallel computation

The worker threads execute the computation, while the calling thread effectively sleeps waiting for the result.

liftPar :: LLVM Native a -> Par Native a

Lift an operation from the base LLVM monad into the Par monad

putIO :: Workers -> Future a -> a -> IO ()

The value represented by a future is now available. Push any blocked continuations to the worker threads.

Orphan instances

Foreign Native

Instance details

Methods

foreignAcc :: (Foreign asm, Typeable a, Typeable b) => asm (a -> b) -> Maybe (a -> Par Native (FutureR Native b))

foreignExp :: (Foreign asm, Typeable x, Typeable y) => asm (x -> y) -> Maybe (IRFun1 Native () (x -> y))