Data.Array.Accelerate.LLVM.Native.Plugin

Description

Documentation

plugin :: Plugin

This GHC plugin is required to support ahead-of-time compilation for the accelerate-llvm-native backend. In particular, it tells GHC about the additional object files generated by runQ* which must be linked into the final executable.

To use it, add the following to the .cabal file of your project:

ghc-options: -fplugin=Data.Array.Accelerate.LLVM.Native.Plugin