Data.Array.Accelerate.LLVM.Native.Plugin

Description

This module provides a GHC plugin to help support ahead-of-time compilation for the accelerate-llvm-native backend. In particular, it tells GHC about the additional object files generated by Accelerate which must be linked into the final executable.

Documentation

plugin :: Plugin

Orphan instances

Serialize Module

Methods

put :: Putter Module

get :: Get Module

Serialize ModuleName

Methods

put :: Putter ModuleName

get :: Get ModuleName

Serialize UnitId

Methods

put :: Putter UnitId

get :: Get UnitId