Data.Array.Accelerate.LLVM.Native.Plugin

Description

This module provides a GHC plugin to help support ahead-of-time compilation for the accelerate-llvm-native backend. In particular, it tells GHC about the additional object files generated by Accelerate which must be linked into the final executable.

Documentation

plugin :: Plugin

data Object

Constructors

Object FilePath